2D TowerDefense Tutorial finished
The project folder TowerDefense is a playable 2D tower defense project, copied from a tutorial on the Coding Quests Youtube channel. There are still some issues with the scaling of all the objects, and there is a lot that needs to be cleaned up, but from here we will transition to the 3D tutorial.